Updated "Plant-Delete"--> only possible to delete if no relationships
* Changed the plant-delete method so that it doesn´t delete in cascade --> Now it is not possible anymore to delete a plant if there is a connection with:
     - chats,
     - buyer, or
     - wishlists.
* Also added the boolean "adIsActive" so that there is a posiblity to deactivate the chat in case that is not possible to delete the chat.

+++ src/main/java/com/sopra/pflanzenkleinanzeigen/controller/PlantController.java
@@ -164,6 +164,12 @@ public class PlantController {
                 model.addAttribute("error", "Sie sind nicht berechtigt, diese Pflanze zu löschen.");
                 return "error";
             }
+            //TODO: unterschiedliche Fehlermeldungen pro case oder möchten wir das lieber in frontend handeln?
+            if(!(plant.getChatsAboutThisPlant().isEmpty() && plant.getWishedBy().isEmpty() && plant.getBuyer() == null)){
+                logger.error("Man darf diese Pflanze nicht löschen, da es chats und/oder Merkzetteln zu diesen Pflanze existieren oder diese Pflanze wurde verkauft.");
+                model.addAttribute("error", "Sie dürfen diese Pflanze nicht löschen, aber sie können die archivieren.");
+                return "error";
+            }
             plantService.deletePlant(plant);
         } catch (Exception deletePlantException) {
             logger.error("Fehler beim Löschen der Pflanze", deletePlantException);
